==> scripts/1 scripts - data cleaning and structuring/1.1 compare 2022 elite checklist and folder.ipynb <==
# import libraries

import pandas as pd
import os

# define globals

path_data_in = "../../data/2022 data/2022 data - 0 raw/NK elite new checklist.xlsx"
path_spreadsheets_folder = "../../data/2022 data/2022 data - 0 raw/NK elite spreadsheets/"
path_data_out = "../../data/2022 data/2022 data - 2 tables/leaders.csv"

# import data: 2022 leaders

dfnew = pd.read_excel(path_data_in).dropna(subset=["LeaderID"])

# rename column names for 2022 leaders

newcolumns = ["LeaderID","Name","NameAlternative","CurrentPositionString","Sex","BirthDateString","BirthPlaceString","DeathReported","DeathString"]
dfnew.columns=newcolumns
dfnew

# import filenames from 2022 leaders spreadsheets

fnames = os.listdir(path_spreadsheets_folder)

fnames

data = [[x.split(" ")[0],x] for x in fnames]

data

dffolder = pd.DataFrame(data,columns=["LeaderID","filename"])

# merge filenames with NK elite master list

dfmerge = dfnew.merge(dffolder,on="LeaderID",how="outer")

dfmerge

# confirm 1-1 match between files and master checklist

dfmerge[dfmerge.filename.isnull()]

dfmerge[dfmerge.Name.isnull()]

# export master checklist as .csv table

dfnew.to_csv(path_data_out,index=False)
