==> scripts/1 scripts - data cleaning and structuring/2 extract data from 2022 leader excel files.ipynb <==
# import libraries

# https://www.python-excel.org/
import xlrd
import pandas as pd
import os

# define global variables

career = pd.DataFrame(columns = ['LeaderID','CareerDateString','CareerString'])
family = pd.DataFrame(columns = ['LeaderID','f1','f2','f3','f4','f5','f6','f7','f8'])
education = pd.DataFrame(columns = ['LeaderID','e1','e2', 'e3', 'e4','e5'])

pathin = "../../data/2022 data/2022 data - 0 raw/NK elite spreadsheets/"
pathout = "../../data/2022 data/2022 data - 2 tables/"

career

len(education.columns)

# define functions

def update_section(pathin, filename, leaderid, df, stopword):
        
    f = xlrd.open_workbook(pathin + filename)
    
    sheet = f.sheet_by_index(0)
    
    dfupdate = pd.DataFrame(columns = df.columns)
        
    for rowindex in range(0,sheet.nrows):
                
        row = sheet.row_values(rowindex)
        
        #print(row)

        # instantiate logical variables
        if(rowindex==0):
            startreadindex = 10000
            readcurrent = False
            readrange = False

        if((rowindex>=startreadindex) and readrange):
            readcurrent = True
            if(row[0]==""):
                #print(row)
                readcurrent = False
                readrange = False

        if(readcurrent):
            #print(row)
            row_to_append = [leaderid]+row[0:len(df.columns)-1]
            #print(row_to_append)
            
            series_to_append = pd.Series(row_to_append,index = df.columns)
            #print(series_to_append)
            
            dfupdate = dfupdate.append(series_to_append,ignore_index=True)
            
            #print(row_to_append)
            #print(series_to_append)
            #print(career_update)

        if(row[0]==stopword):
            startreadindex = rowindex+2
            readrange = True
    
    if(dfupdate.shape[0]==0):
        print("no ",stopword," for: ",leaderid)
        
    return dfupdate

# import excel spreadsheets 

fnames = os.listdir(pathin)

fnames

for filename in fnames:
    leaderid = filename.split(" ")[0]
    
    career_update = update_section(pathin, filename, leaderid, career,"경력사항")
    family_update = update_section(pathin, filename, leaderid, family,"가족사항")
    education_update = update_section(pathin, filename, leaderid, education,"학력사항")
    
    career = career.append(career_update,ignore_index=True).drop_duplicates()
    family = family.append(family_update,ignore_index=True).drop_duplicates()
    education = education.append(education_update,ignore_index=True).drop_duplicates()

# clean tabels and define column names

career

education[education.e3!=""]

education = education.drop(columns=["e3"])

columns_education = ["LeaderID","EducationDateString","EducationSchoolString","EducationMajorString","EducationGraduationStatus"]
education.columns = columns_education
education

education.EducationGraduationStatus.unique()

family

family[family.f5!=""]

family = family.drop(columns=["f5"])

columns_family = ["LeaderID","FamilyMemberRelationship","FamilyMemberName","FamilyMemberSex","FamilyMemberBirthString","FamilyMemberCurrentOrganization","FamilyMemberCurrentPosition","FamilyMemberDeathReported"]
family.columns = columns_family
family

career[career.duplicated()]

len(career["LeaderID"].unique())

605-len(career["LeaderID"].unique())

career[career["LeaderID"]=="김완수a"]

len(family["LeaderID"].unique())

605-len(family["LeaderID"].unique())

family[family.duplicated()]

family[family["LeaderID"]=="강관주"]

education[education.duplicated()]

len(education["LeaderID"].unique())

605-len(education["LeaderID"].unique())

education[education["LeaderID"]=="강관주"]

# export dataframes as .csv tables

career.to_csv(pathout+"career_everything.csv",index=False)

education.to_csv(pathout+"education.csv",index=False)

family.to_csv(pathout+"family.csv",index=False)
