==> scripts/1 scripts - data cleaning and structuring/3.1 family - parse 2015 family relative names for matching.ipynb <==
import pandas as pd
import numpy as np
pd.set_option('display.max_rows', None)

# Step 1 - Import 2015 family data to parse

file_in_raw = "../../data/2015 data/2015 data - 0 raw/leaderfamilylinkmou.csv"

file_out = "../../data/2015 data/2015 data - 0 raw/leaderfamilylinkmou_toparse.xlsx"

file_in_handparsed = "../../data/2015 data/2015 data - 0 raw/leaderfamilylinkmou_toparse_donebyhand.xlsx"

old_family_columns = ["MOUID_2015","Name","FamilyString_2015"]
family2015 = pd.read_csv(file_in_raw,sep=";",dtype="str",names=old_family_columns)

family2015["RelativeName_2015"]=np.nan

# Step 2 - Export 2015 family data and parse by hand

# export and parse by hand
family2015.to_excel(file_out,index=False,encoding="utf-8")

# parse by hand

# Step 3 - Import hand-parsed 2015 family data

# import file parsed by hand
family2015parsed = pd.read_excel(file_in_handparsed,dtype="str")

# check there are no duplicates
family2015parsed[family2015parsed[["Name","RelativeName_2015"]].duplicated()]

# use family2015parsed when merging with 2022 family data
