==> scripts/1 scripts - data cleaning and structuring/3.2 family - merge 2015 and 2022 and parse 9 rows.ipynb <==
# libraries & global vars

import pandas as pd
import numpy as np
pd.set_option('display.max_rows', None)

# import combined leaders

## combined leaders to clean - file in
pathin_combined_leaders_toclean = "../../data/combined data/combined data - 1 cleaning/cleaning step 1 - combined data/leaders_toparse.xlsx"
merged_leaders = pd.read_excel(pathin_combined_leaders_toclean,dtype="str")

# merged_leaders

# path out for combined family table 

## combined family - cleaning - file out
fileout_combined_family_parse = "../../data/combined data/combined data - 1 cleaning/cleaning step 4 - parse strings/toparse_family_9rows.xlsx"

# import and merge 2015 & 2022 family tables

## 2022 family - table - file in
file_2022_family_table = "../../data/2022 data/2022 data - 2 tables/family.xlsx"
new_family = pd.read_excel(file_2022_family_table,dtype="str")

new_family.shape

# 김일성 - 김정숙 is included twice, but these are two different people
new_family[new_family[["LeaderID","RelativeName"]].duplicated(keep=False)]

# 김일성 - 김정숙 is included twice, but these are two different people
new_family[new_family[["LeaderID","RelativeName","RelativeRelationship"]].duplicated()]

## 2015 family - raw - file in
# file_2015_family_raw = "../../data/2015 data/2015 data - 0 raw/leaderfamilylinkmou.csv"
# old_family_columns = ["MOUID_2015","Name","FamilyString_2015"]
# old_family = pd.read_csv(file_2015_family_raw,sep=";",dtype="str",names=old_family_columns)

## 2015 family - hand parsed file in
file_2015_family_parsed = "../../data/2015 data/2015 data - 0 raw/leaderfamilylinkmou_toparse_donebyhand.xlsx"
old_family = pd.read_excel(file_2015_family_parsed,dtype="str")

old_family.rename(columns={"RelativeName_2015":"RelativeName"},inplace=True)

old_family

# add LeaderID to old_family. select columns to merge with new_family
merged_leaders_key = merged_leaders[["LeaderID","MOUID_2015"]]
old_family_withid = old_family.merge(merged_leaders_key,on="MOUID_2015",how="inner")
old_family_withid = old_family_withid[["LeaderID","RelativeName","FamilyString_2015"]]

merged_family = new_family.merge(old_family_withid,on=["LeaderID","RelativeName"],how="outer")

merged_family[merged_family[["LeaderID","RelativeName"]].duplicated(keep=False)]

# family in 2015, but not 2022 - 2 rows
# using file_2015_family_parsed: family in 2015, but not 2022 - 9 rows
merged_family[merged_family.RelativeRelationship.isnull() & merged_family.FamilyString_2015.notnull()].shape

# family in 2022, but not 2015 - 89 rows
# using file_2015_family_parsed: family in 2022, but not 2015 - 134 rows
merged_family[merged_family.RelativeRelationship.notnull() & merged_family.FamilyString_2015.isnull()].shape

# family overlap in 2015 & 2022 - 70 rows
merged_family[merged_family.RelativeRelationship.notnull() & merged_family.FamilyString_2015.notnull()].shape

# rows in file_2015_family_parsed - 78! extra one appears because the DB includes two different 김정숙 related to 김일성
70 + 9

merged_family[merged_family.FamilyString_2015.notnull()]

# rows in 2022 - 204. spot on!
134 + 70

merged_family.shape

# Next steps

### Cleaning Step 2: string disambiguation. Family is already disambiguated.
### Cleaning Step 3: split Career table. Not relevant

### So move on to Cleaning Step 4

### Cleaning Step 4: parse strings. Family has 9 strings from 2015 which aren't parsed.

# sort the 9 rows of 2015only to the top
merged_family.sort_values("RelativeRelationship",na_position="first",inplace=True)

# export to Cleaning Step 4 - to parse 9 rows of 2015only
merged_family.to_excel(fileout_combined_family_parse,index=False,encoding="utf-8")
