==> scripts/1 scripts - data cleaning and structuring/4_merge 2015 & 2022 leader data_export to 9999 - 1 cleaning.ipynb <==
# import libraries and globals

import pandas as pd

# import and merge 2015 & 2022 leader tables

## 2022 leaders - table - file in
file_2022_leaders_table = "../../data/2022 data/2022 data - 2 tables/leaders.csv"

## 2015 leaders - cleaned - file in
file_2015_leaders_cleaned = "../../data/2015 data/2015 data - 1 cleaning/NK elite old checklist.xlsx"

## 9999 leaders - to clean - file out
file_9999_leaders_toclean = "../../data/9999 combined data/9999 data - 1 cleaning/leaders_toparse.csv"

# import 2022 table leaders
new_leaders = pd.read_csv(file_2022_leaders_table)

#list(new_leaders.columns)
new_leaders_change = ["Name","NameAlternative","CurrentPositionString","BirthDateString","BirthPlaceString","DeathString"]
new_leaders_new_columns = [item+"_2022" if item in new_leaders_change else item for item in list(new_leaders.columns)]
new_leaders.columns = new_leaders_new_columns
new_leaders

old_leaders = pd.read_excel(file_2015_leaders_cleaned)

# list(old_leaders.columns)
old_leaders_change = ["MOUID","KoreanName","BirthString","DeathString"]
old_leaders_new_columns = [item+"_2015" if item in old_leaders_change else item for item in list(old_leaders.columns)]
#old_leaders_new_columns
old_leaders.columns = old_leaders_new_columns
old_leaders

# define merged dataframe
#merge_columns = ["LeaderID","Name","Sex","BirthYear","BirthCountry","BirthCP","BirthBCC","DeathReported","BirthString2015","BirthDateString2022","BirthPlaceString2022","CurrentPositionString2022","DeathString"]
#merge_leaders = pd.DataFrame(columns=merge_columns)
merge_leaders = new_leaders.merge(old_leaders,on="LeaderID",how="outer")

merge_leaders

# confirm that names are identical for leaders overlapping in 2015 and 2022
merge_leaders[merge_leaders["KoreanName_2015"].notnull() & merge_leaders["Name_2022"].notnull() & (merge_leaders.KoreanName_2015!=merge_leaders.Name_2022)]

# consolidate Name field
merge_leaders["Name"] = merge_leaders["KoreanName_2015"]
merge_leaders.loc[merge_leaders.KoreanName_2015.isnull(),"Name"] = merge_leaders["Name_2022"]

# rearrange column names
rearranged_column_names = ["LeaderID","Name","Sex","BirthYear","BirthCountry","BirthCP","BirthBCC","DeathReported","MOUID_2015","KoreanName_2015","Name_2022","NameAlternative_2022","BirthString_2015","BirthDateString_2022","BirthPlaceString_2022","DeathString_2015","DeathString_2022"]
merge_leaders = merge_leaders[rearranged_column_names]
merge_leaders

merge_leaders[["Name","KoreanName_2015","Name_2022"]]

# export merged leader table

merge_leaders.to_csv(file_9999_leaders_toclean,index=False)
