==> scripts/1 scripts - data cleaning and structuring/4.1 career - merge and export to combined.ipynb <==
# libraries & global vars

import pandas as pd
import numpy as np

# import combined leaders

## combined leaders to clean - file in
filein_combined_leaders_toclean = "../../data/combined data/combined data - 1 cleaning/cleaning step 1 - combined data/leaders_toparse.xlsx"
merged_leaders = pd.read_excel(filein_combined_leaders_toclean,dtype="str")

# path out for combined career table - to parse and code

## combined career - cleaning - file out
fileout_combined_career_cleaning = "../../data/combined data/combined data - 1 cleaning/cleaning step 1 - combined data/career_everything_toparse.xlsx"

# import and merge 2015 & 2022 career tables

## 2022 career - table - file in
file_2022_career_table = "../../data/2022 data/2022 data - 2 tables/career_everything.xlsx"
new_career = pd.read_excel(file_2022_career_table,dtype="str")
new_career.columns = [item+"_2022" if item != "LeaderID" else item for item in list(new_career.columns)]

## 2015 career - raw - file in
file_2015_career_raw = "../../data/2015 data/2015 data - 0 raw/leaderschoollinkmou.csv"
old_career_columns = ["MOUID_2015", "Name", "CareerStringID_2015", "CareerString_2015", "CareerStartYear", "CareerStartMonth", "EventType", "JointAppointment", "Org1", "Org2", "Org3", "Org4", "Org5", "Position", "OrgLevel", "PositionLevel", "OrgPositionString"]
old_career = pd.read_csv(file_2015_career_raw,sep=";",names=old_career_columns,dtype="str")
old_career = old_career[old_career["MOUID_2015"].notna()]

# add LeaderID to old_family. select columns to merge with new_family
merged_leaders_key = merged_leaders[["LeaderID","MOUID_2015"]]
old_career_withid = old_career.merge(merged_leaders_key,on="MOUID_2015",how="inner")
#old_education_withid = old_education_withid[["LeaderID","FamilyString_2015"]]

merged_career = new_career.merge(old_career_withid,on="LeaderID",how="outer")

merged_career

# career in 2015, but not 2022 - X rows
# merged_education[merged_education.EducationDateString_2022.isnull() & merged_education.EducationString_2015.notnull()]

# career in 2022, but not 2015 - X rows
# merged_education[merged_education.EducationDateString_2022.notnull() & merged_education.EducationString_2015.isnull()]

# career overlap in 2015 & 2022 - X rows
# merged_education[merged_education.EducationDateString_2022.notnull() & merged_education.EducationString_2015.notnull()]

# all null - X rows - can delete these
# merged_education[merged_education.EducationDateString_2022.isnull() & merged_education.EducationString_2015.isnull()]

# export combined education - to cleaning

# export to combined education - to cleaning
merged_career.to_excel(fileout_combined_career_cleaning,index=False,encoding="utf-8")
