==> scripts/1 scripts - data cleaning and structuring/4.2 career - divide career_str into event types.ipynb <==
# libraries & global vars

import pandas as pd
import numpy as np
pd.set_option('display.max_rows', None)

# output path for divided career

pathin_career_undivided = "C:/Users/seoul/Dropbox/00 technical/github/nkelites/data/combined data/combined data - 1 cleaning/cleaning step 4 - career/4.1 career_undivided_unparsed_uncoded/"
pathout_career_divided = "C:/Users/seoul/Dropbox/00 technical/github/nkelites/data/combined data/combined data - 1 cleaning/cleaning step 4 - career/4.2 career_divided_unparsed_uncoded/"

# import career_str

career_str = pd.read_excel(pathin_career_undivided+"career_str_undivided_unparsed_uncoded.xlsx",dtype="str")

career_str.shape

# break out 국가장의위원회 - don't appear to be any joint appts with this

career_str_funeral = career_str[career_str.CareerString.str.contains("국가장의위원회")]

career_str_funeral.shape

# export career_str_funeral to unparsed_uncoded folder
career_str_funeral.to_excel(pathout_career_divided+"career_funeral_unparsed_uncoded.xlsx",index=False,encoding="utf-8")

# extract funeral from remaining career
career_str = career_str[~career_str.CareerString.isin(career_str_funeral.CareerString)]

career_str.shape

# break out education: 유학, 전공, 절업

career_str_education = career_str[career_str.CareerString.str.contains("유학") | career_str.CareerString.str.contains("전공") | career_str.CareerString.str.contains("절업")]

# export career_str_abroad to unparsed_uncoded folder
career_str_education.to_excel(pathout_career_divided+"career_education_unparsed_uncoded.xlsx",index=False,encoding="utf-8")

career_str_education

joint_appt = [140]
non_joint_appt = [item for item in career_str_education.CareerStringID if int(item) not in joint_appt]

# extract (non-joint appointment) funeral from remaining career
career_str = career_str[~career_str.CareerStringID.isin(non_joint_appt)]

career_str.shape

# break out 결혼

career_str_marriage = career_str[career_str.CareerString.str.contains("결혼")]

# export career_str_marriage to unparsed_uncoded folder
career_str_marriage.to_excel(pathout_career_divided+"career_marriage_unparsed_uncoded.xlsx",index=False,encoding="utf-8")

career_str_marriage

joint_appt = [93,3559]
non_joint_appt = [item for item in career_str_marriage.CareerStringID if int(item) not in joint_appt]

# extract (non-joint appointment) marriage from remaining career
career_str = career_str[~career_str.CareerStringID.isin(non_joint_appt)]

career_str.shape

# break out 수훈 & 칭호

# 수훈 및 칭호
career_str_award = career_str[career_str.CareerString.str.contains("수훈") | career_str.CareerString.str.contains("칭호")]

# export career_str_marriage to unparsed_uncoded folder
career_str_award.to_excel(pathout_career_divided+"career_award_unparsed_uncoded.xlsx",index=False,encoding="utf-8")

career_str_award.shape

joint_appt = [407, 487, 691, 773, 778, 785, 1403, 1404, 1405, 1406, 1407, 1415, 1418, 1419, 1425, 1427, 1428, 3627, 3690, 4872, 5725, 6026, 6486, 6507, 6551, 6653]
non_joint_appt = [item for item in career_str_award.CareerStringID if int(item) not in joint_appt]

# extract (non-joint appointment) marriage from remaining career
career_str = career_str[~career_str.CareerStringID.isin(non_joint_appt)]

career_str.shape

# break out 사망, 사형, 처형

# 사형, 사망, 처형
career_str_death = career_str[career_str.CareerString.str.contains("처형") | career_str.CareerString.str.contains("사형") | career_str.CareerString.str.contains("사망")]

# export career_str_marriage to unparsed_uncoded folder
career_str_death.to_excel(pathout_career_divided+"career_death_unparsed_uncoded.xlsx",index=False,encoding="utf-8")

career_str_death.shape

career_str = career_str[~career_str.CareerStringID.isin(career_str_death.CareerStringID)]

career_str.shape

# break out visits: 동행, 참석시, 습격, 방문,  빨치산 활동

career_str_visit = career_str[career_str.CareerString.str.contains(" 동행") | career_str.CareerString.str.contains("참석시") | career_str.CareerString.str.contains("습격") | career_str.CareerString.str.contains("방문") | career_str.CareerString.str.contains("빨치산 활동")]

career_str_visit

# export career_str_marriage to unparsed_uncoded folder
career_str_visit.to_excel(pathout_career_divided+"career_visit_unparsed_uncoded.xlsx",index=False,encoding="utf-8")

career_str_visit.shape

career_str_visit

# extract (non-joint appointment) marriage from remaining career
career_str = career_str[~career_str.CareerStringID.isin(career_str_visit.CareerStringID)]

career_str.shape

# export career_str jobs remaining after divisions

career_str.to_excel(pathout_career_divided+"career_jobs_unparsed_uncoded.xlsx",index=False,encoding="utf-8")
