==> scripts/1 scripts - data cleaning and structuring/99 organization tree.ipynb <==
import pandas as pd
import numpy as np

career_2015_old_path = "../../data/2015 data/2015 data - 0 raw/"
career_2015_old_file = "leadercareerlinkmou.csv"
career_2015_old_header = ["MOUID","KoreanName","CareerStringID","CareerString","CareerStartYear","CareerStartMonth","EventType","JointAppointment","Org1","Org2","Org3","Org4","Org5","Position","OrgLevel","PositionLevel","OrgPositionString"]
career_2015_old = pd.read_csv(career_2015_old_path+career_2015_old_file,sep=";",dtype="str",names=career_2015_old_header)

career_2015_cleaning_path = "../../data/2015 data/2015 data - 1 cleaning/"
career_2015_cleaning_file = "Eunhos latest_CareerStringsCleaned.xlsx"
career_2015_cleaning = pd.read_excel(career_2015_cleaning_path+career_2015_cleaning_file,dtype="str")

career_2022_path = "../../data/2022 data/2022 data - 2 tables/"
career_2022_file = "career_everything.csv"
career_2022 = pd.read_csv(career_2022_path+career_2022_file,dtype="str")

career_2022
career_2015_cleaning
#career_2015_old

tree_cleaning = career_2015_cleaning[["Org1","Org2","Org3","Org4"]].drop_duplicates()

tree_cleaning["Org5"]=np.NaN
tree_cleaning

tree_old = career_2015_old[["Org1","Org2","Org3","Org4","Org5"]].drop_duplicates()

tree_concat = pd.concat([tree_old,tree_cleaning])

tree_concat = tree_concat.drop_duplicates()

tree_concat.sort_values(["Org1","Org2","Org3","Org4","Org5"],inplace=True)

tree_concat

tree_concat.to_excel("../../data/combined data/combined data - 3 queries/tree.xlsx",index=False,encoding="utf-8")
