==> scripts/1 scripts - data cleaning and structuring/99_extract uncoded career strings from old data.ipynb <==
import pandas as pd

## Extract uncoded career strings from old data 

columnnames=["MOUID", "KoreanName", "CareerStringID", "CareerString", "CareerStartYear", "CareerStartMonth", "EventType", "JointAppointment", "Org1", "Org2", "Org3", "Org4", "Org5", "Position", "OrgLevel", "PositionLevel", "OrgPositionString"]
df=pd.read_csv("../old data/leader-career link table_dsl.csv",names=columnnames,header=None,delimiter=";",encoding="utf-8")

df

# No. of CareerString with no Org1 = 3076
# No. of UNIQUE CareerString with no Org1 = 2148
q1 = df[df["Org1"].isnull()][["CareerStringID","CareerString"]]
q1.shape
q1.drop_duplicates().shape

# No. of CareerString with no CareerStartYear = 3076
# No. of UNIQUE CareerString with no CareerStartYear = 2148
q1 = df[df["CareerStartYear"].isnull()][["CareerStringID","CareerString"]]
q1.shape
q1.drop_duplicates().shape

q1unique=q1.drop_duplicates()
q1unique

q1unique.to_excel("../old data/working files/uncoded careerstrings_old data.xlsx",encoding="utf-8")

## Organization Tree, for reference when coding

dforgcounts=dforg.groupby(["Org1","Org2","Org3","Org4","Org5"],dropna=False).size().reset_index(name='counts')
dforgcounts=dforgcounts.sort_values(by=["Org1","Org2","Org3","Org4","Org5","counts"])
dforgcounts

dforgcounts.to_excel("../old data/working files/organization tree.xlsx",encoding="utf-8")
