==> scripts/1 scripts - data cleaning and structuring/99_extract uncoded strings from new leaders.ipynb <==
# import libraries & define globals

import pandas as pd

path_2015_leaders = "../../data/2015 data/2015 data - 1 cleaning/NK elite old checklist.xlsx"
path_2022_leaders = "../../data/2022 data/2022 data - 2 tables/leaders.csv"

# import 2015 and 2022 data 

# dfold = pd.read_csv("../data/old data/leader table_dsl.csv",delimiter=";",encoding="utf-8")
dfold = pd.read_excel(path_2015_leaders).dropna(subset=["LeaderID"])

dfnew = pd.read_csv(path_2022_leaders).dropna(subset=["LeaderID"])

dfnew

dfold

newdupa = dfnew[dfnew["LeaderID"].str.contains("a")]

newdupnames = newdupa.Name.tolist()
len(newdupnames)

newdupnames

olddupa = dfold[dfold["LeaderID"].str.contains("a")]

olddupa

dfold[dfold.KoreanName.str.contains("김경호")]

dfnew[dfnew.Name.str.contains("김경호")]

dfnew

# merge old and new - confusion matrix

### confirm unique LeaderID in new and old data

dfold.shape

len(dfold.LeaderID.unique())

dfold[dfold.LeaderID.duplicated()==True]

dfold[dfold.LeaderID.str.contains("김완수")]

dfnew[dfnew.LeaderID.str.contains("김완수")]

dfnew.shape

len(dfnew.LeaderID.unique())

### merge new and old leader lists

dfmerge = dfnew.merge(dfold,on="LeaderID",how="outer")
dfmerge.shape

dfmerge

#dfmerge[dfmerge.LeaderID.str.contains("a")]
len(dfmerge.LeaderID.unique())

### confusion matrix 

# contained in new, but not in old - cases where MOUID is null
dfmerge[dfmerge.MOUID.isnull()].shape

# contained in old, but not in new - cases where Name is null
dfmerge[dfmerge.Name.isnull()].shape

# contained in both old and new - cases where neither Name nor MOUID are null
dfmerge[dfmerge.Name.notnull() & dfmerge.MOUID.notnull()].shape

# contained in neither old nor new - cases where both Name and MOUID are null
dfmerge[dfmerge.Name.isnull() & dfmerge.MOUID.isnull()].shape

# confirm the intersection and unique new sum to total old (384)
321+63

# confirm the intersection and unique old sum to total new (605)
321+284

# confirm the three cells sum to the total (668)
284+63+321

# examine parsed fields for old and new - e.g. BirthString & 생년월일

dfmerge

dfmerge.columns = ['LeaderID', 'New_Name', 'New_Name2', 'New_Position', 'New_Sex', 'New_BirthString', 'New_BirthPlaceString', 'New_DeathStatus','New_DeathString', 'Old_MOUID', 'Old_Name', 'Old_BirthString', 'Old_DeathString','BirthYear', 'BirthCountry', 'BirthCP', 'BirthBCC']

dfmerge

dfmerge[dfmerge.New_Name.notnull() & dfmerge.Old_MOUID.notnull()]

dfmerge[dfmerge.LeaderID.str.contains("김완수")]

dfmerge.New_Sex.unique()

dfmerge["DeathStatus"]=""
dfmerge["DeathYear"]=""
dfmerge["Imputed_Sex"]=dfmerge["New_Sex"]
dfmerge["InconsistentBirthDate"]=""
dfmerge["InconsistentBirthPlace"]=""
dfmerge["InconsistentDeathString"]=""

dfmerge.columns

dfmerge.Imputed_Sex.unique()

# data subsets for parsing & imputing

# new, not old
dfmerge_new_noold = dfmerge[dfmerge.Old_MOUID.isnull()]
dfmerge_new_noold = dfmerge_new_noold[['LeaderID','New_BirthString', 'BirthYear', 'New_BirthPlaceString','BirthCountry', 'BirthCP', 'BirthBCC','New_DeathStatus','New_DeathString',"DeathStatus","DeathYear"]]
dfmerge_new_noold.to_excel("../data cleaning/2022.05 clean leader table/merge_new_noold.xlsx",index=False)

dfmerge.columns

# new and old
dfmerge_new_old = dfmerge[dfmerge.Old_MOUID.notnull() & dfmerge.New_Name.notnull()]
dfmerge_new_old = dfmerge_new_old[['LeaderID',"Imputed_Sex",'New_BirthString','Old_BirthString','BirthYear','InconsistentBirthDate', 'New_BirthPlaceString',"Old_BirthString",'BirthCountry', 'BirthCP', 'BirthBCC',"InconsistentBirthPlace",'New_DeathStatus','New_DeathString',"Old_DeathString","DeathStatus","DeathYear","InconsistentDeathString"]]
dfmerge_new_old.to_excel("../data cleaning/2022.05 clean leader table/merge_new_old.xlsx",index=False)
